refactor: enhance ReferFriends navigation logic
- Introduced new hooks `useReplaceToReferFriends` for both native and web platforms to streamline navigation based on login status.
- Updated `BreadcrumbSection` and `useRedirectWhenNotLoggedIn` to utilize the new navigation hooks.
- Removed the obsolete `useNavigateToReferFriends` hook to improve code clarity and maintainability.
